Update MotorcycleStock model to use BigDecimal rather than Float to avoid rounding imprecisions

+++ src/main/java/com/ibardos/motoShop/entity/MotorcycleStock.java
@@ -26,7 +26,7 @@ public class MotorcycleStock {
     private MotorcycleModel motorcycleModel;
     private int mileage;
     private BigDecimal purchasingPrice;
-    private Float profitMargin;
+    private BigDecimal profitMargin;
     private BigDecimal profitOnUnit;
     private BigDecimal sellingPrice;
     private int inStock;
